pyproject.toml
[project]
name = "credit-invoice-summariser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "tedlundwall", email = "tedlundwall@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "polars[pandas]>=1.38.1",
    "tabula-py>=2.10.0",
]

[project.scripts]
credit-invoice-summariser = "credit_invoice_summariser:main"

[build-system]
requires = ["uv_build>=0.10.9,<0.11.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pyinstaller>=6.19.0",
]
